Move go.mod. Add extern out logic
