+++ bindings/go/entities/address.go
@@ -6,7 +6,7 @@ import (
 	"errors"
 	"io"
 
-	veriblock "github.com/VeriBlock/alt-integration-cpp"
+	veriblock "github.com/VeriBlock/alt-integration-cpp/bindings/go"
 )
 
 // AddressType ...
