+++ bindings/go/entities/alt_block_addon.go
@@ -4,7 +4,7 @@ import (
 	"io"
 	"math"
 
-	veriblock "github.com/VeriBlock/alt-integration-cpp"
+	veriblock "github.com/VeriBlock/alt-integration-cpp/bindings/go"
 )
 
 // AltBlockAddon ...
