+++ bindings/go/entities/alt_endorsement.go
@@ -3,7 +3,7 @@ package entities
 import (
 	"io"
 
-	veriblock "github.com/VeriBlock/alt-integration-cpp"
+	veriblock "github.com/VeriBlock/alt-integration-cpp/bindings/go"
 )
 
 // AltEndorsement ...
