+++ bindings/go/entities/btcblock.go
@@ -6,7 +6,7 @@ import (
 	"encoding/binary"
 	"io"
 
-	veriblock "github.com/VeriBlock/alt-integration-cpp"
+	veriblock "github.com/VeriBlock/alt-integration-cpp/bindings/go"
 )
 
 // BtcBlock ...
