+++ bindings/go/entities/popdata.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 	"io"
 
-	veriblock "github.com/VeriBlock/alt-integration-cpp"
+	veriblock "github.com/VeriBlock/alt-integration-cpp/bindings/go"
 )
 
 // PopData ...
