+++ bindings/go/entities/vbk_block_addon.go
@@ -5,7 +5,7 @@ import (
 	"io"
 	"math"
 
-	veriblock "github.com/VeriBlock/alt-integration-cpp"
+	veriblock "github.com/VeriBlock/alt-integration-cpp/bindings/go"
 )
 
 // VbkBlockAddon ...
