+++ bindings/go/entities/vbktx.go
@@ -5,7 +5,7 @@ import (
 	"encoding/binary"
 	"io"
 
-	veriblock "github.com/VeriBlock/alt-integration-cpp"
+	veriblock "github.com/VeriBlock/alt-integration-cpp/bindings/go"
 )
 
 // VbkTx ...
