+++ bindings/go/ffi/pop_context.go
@@ -10,7 +10,7 @@ import "unsafe"
 var (
 	OnGetAltchainID      = func() int { panic("OnGetAltchainID not set") }
 	OnGetBootstrapBlock  = func() string { panic("OnGetBootstrapBlock not set") }
-	OnGetBlockHeaderHash = func() { panic("OnGetBlockHeaderHash not set") }
+	OnGetBlockHeaderHash = func(toBeHashed []byte) []byte { panic("OnGetBlockHeaderHash not set") }
 	OnGetAtv             = func() { panic("OnGetAtv not set") }
 	OnGetVtb             = func() { panic("OnGetVtb not set") }
 	OnGetVbk             = func() { panic("OnGetVbk not set") }
